Move RPC tests to lib module
